Add files via upload
This is an updated version of Account_Saver. In this version i've added new commands which will let you to view how many accounts you have saved without opening the account file, check if an account already exists and add multiple files with the GUI without you having to to reopen the GUI or clear each column to add new account.